dtype to device update
First of all, thank you for this great library. I was trying to use the FocalLoss and I found a problem that always give an error due to expect a source to be on CUDA instead of CPU. Anyway I belive you have a typo on line 86, where to(input.dtype)) should be to(input.device)). I copied you source code and by changing this line I manage to make it work perfectly.