Fix suppression of unwanted device/dtype test combos
The previous method for doing this was based around always requiring
the device fixture when the dtype fixture was used, and then skipping
the test from the dtype fixture when an unwanted combo is detected.
This can cause tests which require dtype but not device to run too
many times.