fix: missed single pixel instance in upper left corner
# Avoid missing a single pixel instance in the upper left corner

Example of issue:
```
target = torch.tensor(
    [
        [1, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
        [1, 0, 0, 0, 0],
        [0, 0, 0, 0, 0],
        [1, 0, 1, 0, 1],
    ]
).view(1, 1, 5, 5)
out = connected_components(target)
```
-->
BEFORE:
```
tensor([[[[ 0,  0,  0,  0,  0],
          [ 0,  0,  0,  0,  0],
          [10,  0,  0,  0,  0],
          [ 0,  0,  0,  0,  0],
          [22,  0, 24,  0, 26]]]])
```

PROPOSED:
-->
```
tensor([[[[ 1,  0,  0,  0,  0],
          [ 0,  0,  0,  0,  0],
          [11,  0,  0,  0,  0],
          [ 0,  0,  0,  0,  0],
          [21,  0, 23,  0, 25]]]])
```
--- 

The algorithm starts by placing unique values for every pixel in the foreground (with torch.arange) and zeroing out the background with a mask. The value of the first pixel is then 0 by default which is confused for background in the spacial case above. The proposed solution is just to add one to all initial values.