Fix homography between unit square and scaled version (#3335)
* testcase for points on square

* Fixes the case of 2D points being the vertices of unit square

When the number of correspondences (x, x') is exactly 4,this fix changes the homogenous linear system of equations from Ah=0 to Ah=x'. This is becasue LU decomposition expects a squared matix, which currently is computed as A.T@A. However, in the case of x being the vertices of the unit square and x' a scaled version A.T@A is singular an has no inverse. The resulting h has NaN values.

When there are exactly 4 correspondences, the A in Ah=x' is squared and non-singular and h is well-behaved. Since it the homography is up-to-scale we set h33=1